+++ src/test/java/com/github/victools/jsonschema/generator/impl/SchemaGenerationContextImplTest.java
@@ -17,9 +17,18 @@
 package com.github.victools.jsonschema.generator.impl;
 
 import com.fasterxml.classmate.ResolvedType;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.github.victools.jsonschema.generator.AbstractTypeAwareTest;
+import com.github.victools.jsonschema.generator.CustomDefinition;
 import com.github.victools.jsonschema.generator.CustomDefinitionProviderV2;
+import com.github.victools.jsonschema.generator.CustomPropertyDefinition;
+import com.github.victools.jsonschema.generator.FieldScope;
+import com.github.victools.jsonschema.generator.MethodScope;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfig;
+import com.github.victools.jsonschema.generator.SchemaKeyword;
+import com.github.victools.jsonschema.generator.SchemaVersion;
 import java.util.Arrays;
 import java.util.Collections;
 import org.junit.Assert;
@@ -30,27 +39,60 @@ import org.mockito.Mockito;
 /**
  * Test for the {@link SchemaGenerationContextImpl} class.
  */
-public class SchemaGenerationContextImplTest {
+public class SchemaGenerationContextImplTest extends AbstractTypeAwareTest {
 
-    private SchemaGenerationContextImpl context;
+    private SchemaGenerationContextImpl contextImpl;
+
+    public SchemaGenerationContextImplTest() {
+        super(TestClass.class);
+    }
 
     @Before
     public void setUp() {
+        SchemaVersion schemaVersion = SchemaVersion.DRAFT_2019_09;
+        this.prepareContextForVersion(schemaVersion);
+
         SchemaGeneratorConfig config = Mockito.mock(SchemaGeneratorConfig.class);
-        this.context = new SchemaGenerationContextImpl(config, TypeContextFactory.createDefaultTypeContext());
+        Mockito.when(config.getSchemaVersion()).thenReturn(schemaVersion);
+        Mockito.when(config.getKeyword(Mockito.any()))
+                .thenAnswer(invocation -> ((SchemaKeyword) invocation.getArgument(0)).forVersion(schemaVersion));
+        Mockito.when(config.resolveTitle(Mockito.any(FieldScope.class))).thenReturn("Field Title");
+        Mockito.when(config.resolveTitle(Mockito.any(MethodScope.class))).thenReturn("Method Title");
+        Mockito.when(config.resolveDescriptionForType(Mockito.any())).thenReturn("Type Description");
+        Mockito.when(config.resolveStringMinLength(Mockito.any(FieldScope.class))).thenReturn(null);
+        Mockito.when(config.resolveStringMaxLength(Mockito.any(FieldScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayMinItems(Mockito.any(FieldScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayMaxItems(Mockito.any(FieldScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayUniqueItems(Mockito.any(FieldScope.class))).thenReturn(null);
+        Mockito.when(config.resolveStringMinLength(Mockito.any(MethodScope.class))).thenReturn(null);
+        Mockito.when(config.resolveStringMaxLength(Mockito.any(MethodScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayMinItems(Mockito.any(MethodScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayMaxItems(Mockito.any(MethodScope.class))).thenReturn(null);
+        Mockito.when(config.resolveArrayUniqueItems(Mockito.any(MethodScope.class))).thenReturn(null);
+        Mockito.when(config.resolveStringMinLengthForType(Mockito.any())).thenReturn(null);
+        Mockito.when(config.resolveStringMaxLengthForType(Mockito.any())).thenReturn(null);
+        Mockito.when(config.resolveArrayMinItemsForType(Mockito.any())).thenReturn(null);
+        Mockito.when(config.resolveArrayMaxItemsForType(Mockito.any())).thenReturn(null);
+        Mockito.when(config.resolveArrayUniqueItemsForType(Mockito.any())).thenReturn(null);
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        Mockito.when(config.createObjectNode()).then(_invocation -> objectMapper.createObjectNode());
+        Mockito.when(config.createArrayNode()).then(_invocation -> objectMapper.createArrayNode());
+
+        this.contextImpl = new SchemaGenerationContextImpl(config, this.getContext().getTypeContext());
     }
 
     @Test
     public void testHandlingDefinition_ExistentNode() {
         ResolvedType javaType = Mockito.mock(ResolvedType.class);
         ObjectNode definitionInput = Mockito.mock(ObjectNode.class);
-        SchemaGenerationContextImpl returnValue = this.context.putDefinition(javaType, definitionInput, null);
+        SchemaGenerationContextImpl returnValue = this.contextImpl.putDefinition(javaType, definitionInput, null);
 
-        Assert.assertSame(this.context, returnValue);
-        Assert.assertTrue(this.context.containsDefinition(javaType, null));
+        Assert.assertSame(this.contextImpl, returnValue);
+        Assert.assertTrue(this.contextImpl.containsDefinition(javaType, null));
         DefinitionKey key = new DefinitionKey(javaType, null);
-        Assert.assertSame(definitionInput, this.context.getDefinition(key));
-        Assert.assertEquals(Collections.singleton(key), this.context.getDefinedTypes());
+        Assert.assertSame(definitionInput, this.contextImpl.getDefinition(key));
+        Assert.assertEquals(Collections.singleton(key), this.contextImpl.getDefinedTypes());
     }
 
     @Test
@@ -58,9 +100,9 @@ public class SchemaGenerationContextImplTest {
         ResolvedType javaType = Mockito.mock(ResolvedType.class);
         DefinitionKey key = new DefinitionKey(javaType, null);
 
-        Assert.assertFalse(this.context.containsDefinition(javaType, null));
-        Assert.assertNull(this.context.getDefinition(key));
-        Assert.assertEquals(Collections.<DefinitionKey>emptySet(), this.context.getDefinedTypes());
+        Assert.assertFalse(this.contextImpl.containsDefinition(javaType, null));
+        Assert.assertNull(this.contextImpl.getDefinition(key));
+        Assert.assertEquals(Collections.<DefinitionKey>emptySet(), this.contextImpl.getDefinedTypes());
     }
 
     @Test
@@ -69,29 +111,29 @@ public class SchemaGenerationContextImplTest {
         DefinitionKey key = new DefinitionKey(javaType, null);
 
         // initially, all lists are empty
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getReferences(key));
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getNullableReferences(key));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getReferences(key));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getNullableReferences(key));
 
         // adding a not-nullable entry creates the "references" list
         ObjectNode referenceInputOne = Mockito.mock(ObjectNode.class);
-        SchemaGenerationContextImpl returnValue = this.context.addReference(javaType, referenceInputOne, null, false);
-        Assert.assertSame(this.context, returnValue);
-        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.context.getReferences(key));
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getNullableReferences(key));
+        SchemaGenerationContextImpl returnValue = this.contextImpl.addReference(javaType, referenceInputOne, null, false);
+        Assert.assertSame(this.contextImpl, returnValue);
+        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.contextImpl.getReferences(key));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getNullableReferences(key));
 
         // adding another not-nullable entry adds it to the existing "references" list
         ObjectNode referenceInputTwo = Mockito.mock(ObjectNode.class);
-        returnValue = this.context.addReference(javaType, referenceInputTwo, null, false);
-        Assert.assertSame(this.context, returnValue);
-        Assert.assertEquals(Arrays.asList(referenceInputOne, referenceInputTwo), this.context.getReferences(key));
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getNullableReferences(key));
+        returnValue = this.contextImpl.addReference(javaType, referenceInputTwo, null, false);
+        Assert.assertSame(this.contextImpl, returnValue);
+        Assert.assertEquals(Arrays.asList(referenceInputOne, referenceInputTwo), this.contextImpl.getReferences(key));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getNullableReferences(key));
 
         // adding a nullable entry creates the "nullableReferences" list
         ObjectNode referenceInputThree = Mockito.mock(ObjectNode.class);
-        returnValue = this.context.addReference(javaType, referenceInputThree, null, true);
-        Assert.assertSame(this.context, returnValue);
-        Assert.assertEquals(Arrays.asList(referenceInputOne, referenceInputTwo), this.context.getReferences(key));
-        Assert.assertEquals(Collections.singletonList(referenceInputThree), this.context.getNullableReferences(key));
+        returnValue = this.contextImpl.addReference(javaType, referenceInputThree, null, true);
+        Assert.assertSame(this.contextImpl, returnValue);
+        Assert.assertEquals(Arrays.asList(referenceInputOne, referenceInputTwo), this.contextImpl.getReferences(key));
+        Assert.assertEquals(Collections.singletonList(referenceInputThree), this.contextImpl.getNullableReferences(key));
     }
 
     @Test
@@ -103,15 +145,15 @@ public class SchemaGenerationContextImplTest {
 
         // adding an entry creates the "references" list for that type
         ObjectNode referenceInputOne = Mockito.mock(ObjectNode.class);
-        this.context.addReference(javaTypeOne, referenceInputOne, null, false);
-        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.context.getReferences(keyOne));
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getReferences(keyTwo));
+        this.contextImpl.addReference(javaTypeOne, referenceInputOne, null, false);
+        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.contextImpl.getReferences(keyOne));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getReferences(keyTwo));
 
         // adding an entry for another type creates a separate "references" list for this other type
         ObjectNode referenceInputTwo = Mockito.mock(ObjectNode.class);
-        this.context.addReference(javaTypeTwo, referenceInputTwo, null, false);
-        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.context.getReferences(keyOne));
-        Assert.assertEquals(Collections.singletonList(referenceInputTwo), this.context.getReferences(keyTwo));
+        this.contextImpl.addReference(javaTypeTwo, referenceInputTwo, null, false);
+        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.contextImpl.getReferences(keyOne));
+        Assert.assertEquals(Collections.singletonList(referenceInputTwo), this.contextImpl.getReferences(keyTwo));
     }
 
     @Test
@@ -124,14 +166,127 @@ public class SchemaGenerationContextImplTest {
 
         // adding an entry creates the "references" list for that type
         ObjectNode referenceInputOne = Mockito.mock(ObjectNode.class);
-        this.context.addReference(javaType, referenceInputOne, providerOne, false);
-        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.context.getReferences(keyOne));
-        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.context.getReferences(keyTwo));
+        this.contextImpl.addReference(javaType, referenceInputOne, providerOne, false);
+        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.contextImpl.getReferences(keyOne));
+        Assert.assertEquals(Collections.<ObjectNode>emptyList(), this.contextImpl.getReferences(keyTwo));
 
         // adding an entry for the same type but other ignored definition provider creates a separate "references" list for this other type
         ObjectNode referenceInputTwo = Mockito.mock(ObjectNode.class);
-        this.context.addReference(javaType, referenceInputTwo, providerTwo, false);
-        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.context.getReferences(keyOne));
-        Assert.assertEquals(Collections.singletonList(referenceInputTwo), this.context.getReferences(keyTwo));
+        this.contextImpl.addReference(javaType, referenceInputTwo, providerTwo, false);
+        Assert.assertEquals(Collections.singletonList(referenceInputOne), this.contextImpl.getReferences(keyOne));
+        Assert.assertEquals(Collections.singletonList(referenceInputTwo), this.contextImpl.getReferences(keyTwo));
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_noCustomDefinition() {
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"allOf\":[{},{\"title\":\"Field Title\"}]}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_withCustomTypeDefinition() {
+        Mockito.doReturn(new CustomDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom type"), true))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(ResolvedType.class), Mockito.any(), Mockito.any());
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"allOf\":[{\"$comment\":\"custom type\",\"description\":\"Type Description\"},{\"title\":\"Field Title\"}]}",
+                result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_withCustomPropertyDefinition() {
+        Mockito.doReturn(new CustomPropertyDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom property")))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(FieldScope.class), Mockito.any(), Mockito.any());
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"$comment\":\"custom property\",\"title\":\"Field Title\"}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_withCustomPropertyAndTypeDefinitions() {
+        Mockito.doReturn(new CustomDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom type"), true))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(ResolvedType.class), Mockito.any(), Mockito.any());
+        Mockito.doReturn(new CustomPropertyDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom property")))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(FieldScope.class), Mockito.any(), Mockito.any());
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"$comment\":\"custom property\",\"title\":\"Field Title\"}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForMethod_noCustomDefinition() {
+        MethodScope targetMethod = this.getTestClassMethod("isBooleanField");
+        JsonNode result = this.contextImpl.createStandardDefinitionReference(targetMethod, null);
+        Assert.assertEquals("{\"allOf\":[{},{\"title\":\"Method Title\"}]}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForMethod_withCustomTypeDefinition() {
+        Mockito.doReturn(new CustomDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom type"), true))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(ResolvedType.class), Mockito.any(), Mockito.any());
+        MethodScope targetMethod = this.getTestClassMethod("isBooleanField");
+        JsonNode result = this.contextImpl.createStandardDefinitionReference(targetMethod, null);
+        Assert.assertEquals("{\"allOf\":[{\"$comment\":\"custom type\",\"description\":\"Type Description\"},{\"title\":\"Method Title\"}]}",
+                result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForMethod_withCustomPropertyDefinition() {
+        Mockito.doReturn(new CustomPropertyDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom property")))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(MethodScope.class), Mockito.any(), Mockito.any());
+        MethodScope targetMethod = this.getTestClassMethod("isBooleanField");
+        JsonNode result = this.contextImpl.createStandardDefinitionReference(targetMethod, null);
+        Assert.assertEquals("{\"$comment\":\"custom property\",\"title\":\"Method Title\"}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForMethod_withCustomPropertyAndTypeDefinitions() {
+        Mockito.doReturn(new CustomDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom type"), true))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(ResolvedType.class), Mockito.any(), Mockito.any());
+        Mockito.doReturn(new CustomPropertyDefinition(this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom property")))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(MethodScope.class), Mockito.any(), Mockito.any());
+        MethodScope targetMethod = this.getTestClassMethod("isBooleanField");
+        JsonNode result = this.contextImpl.createStandardDefinitionReference(targetMethod, null);
+        Assert.assertEquals("{\"$comment\":\"custom property\",\"title\":\"Method Title\"}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_withCustomTypeDefinitionExcludingAttributes() {
+        ObjectNode customNode = this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom type");
+        Mockito.doReturn(new CustomDefinition(customNode, CustomDefinition.INLINE_DEFINITION, CustomDefinition.EXCLUDING_ATTRIBUTES))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(ResolvedType.class), Mockito.any(), Mockito.any());
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"allOf\":[{\"$comment\":\"custom type\"},{\"title\":\"Field Title\"}]}", result.toString());
+    }
+
+    @Test
+    public void testCreateStandardDefinitionReferenceForField_withCustomPropertyDefinitionExcludingAttributes() {
+        ObjectNode customNode = this.contextImpl.getGeneratorConfig().createObjectNode().put("$comment", "custom property");
+        Mockito.doReturn(new CustomPropertyDefinition(customNode, CustomDefinition.EXCLUDING_ATTRIBUTES))
+                .when(this.contextImpl.getGeneratorConfig())
+                .getCustomDefinition(Mockito.any(FieldScope.class), Mockito.any(), Mockito.any());
+        FieldScope targetField = this.getTestClassField("booleanField");
+        ObjectNode result = this.contextImpl.createStandardDefinitionReference(targetField, null);
+        Assert.assertEquals("{\"$comment\":\"custom property\"}", result.toString());
+    }
+
+    private static class TestClass {
+
+        public boolean booleanField;
+
+        public boolean isBooleanField() {
+            return this.booleanField;
+        }
     }
 }